Get_mat for MatZq
