+++ src/integer_mod_q/mat_zq/get.rs
@@ -18,6 +18,7 @@ use crate::{
 };
 use flint_sys::{
     fmpz::{fmpz, fmpz_set},
+    fmpz_mat::fmpz_mat_set,
     fmpz_mod_mat::{
         fmpz_mod_mat_entry, fmpz_mod_mat_init_set, fmpz_mod_mat_window_clear,
         fmpz_mod_mat_window_init,
@@ -26,6 +27,30 @@ use flint_sys::{
 use std::{fmt::Display, mem::MaybeUninit};
 
 impl MatZq {
+    /// Create a [`MatZ`] from a [`MatZq`].
+    ///
+    /// The output matrix entries values will be in the range `[0, Modulus)`.
+    /// Use [`MatZq::get_closest_to_zero_representative`] if they should be
+    /// close to `0`.
+    ///
+    /// Returns the matrix as a [`MatZ`].
+    ///
+    /// # Examples
+    /// ```
+    /// use qfall_math::integer::MatZ;
+    /// use qfall_math::integer_mod_q::MatZq;
+    /// use std::str::FromStr;
+    ///
+    /// let m = MatZq::from_str("[[1, 2],[3, -1]] mod 5").unwrap();
+    ///
+    /// let a = m.get_mat();
+    /// ```
+    pub fn get_mat(&self) -> MatZ {
+        let mut out = MatZ::new(self.get_num_rows(), self.get_num_columns());
+        unsafe { fmpz_mat_set(&mut out.matrix, &self.matrix.mat[0]) };
+        out
+    }
+
     /// Returns the modulus of the matrix as a [`Modulus`].
     ///
     /// # Examples
@@ -38,127 +63,7 @@ impl MatZq {
     pub fn get_mod(&self) -> Modulus {
         self.modulus.clone()
     }
-}
 
-impl GetNumRows for MatZq {
-    /// Returns the number of rows of the matrix as a [`i64`].
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::integer_mod_q::MatZq;
-    /// use qfall_math::traits::*;
-    ///
-    /// let matrix = MatZq::new(5, 6, 7);
-    /// let rows = matrix.get_num_rows();
-    /// ```
-    fn get_num_rows(&self) -> i64 {
-        self.matrix.mat[0].r
-    }
-}
-
-impl GetNumColumns for MatZq {
-    /// Returns the number of columns of the matrix as a [`i64`].
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::integer_mod_q::MatZq;
-    /// use qfall_math::traits::*;
-    ///
-    /// let matrix = MatZq::new(5, 6, 7);
-    /// let rows = matrix.get_num_columns();
-    /// ```
-    fn get_num_columns(&self) -> i64 {
-        self.matrix.mat[0].c
-    }
-}
-
-impl GetEntry<Z> for MatZq {
-    /// Outputs the [`Z`] value of a specific matrix entry.
-    ///
-    /// Parameters:
-    /// - `row`: specifies the row in which the entry is located
-    /// - `column`: specifies the column in which the entry is located
-    ///
-    /// Negative indices can be used to index from the back, e.g., `-1` for
-    /// the last element.
-    ///
-    /// Returns the [`Z`] value of the matrix at the position of the given
-    /// row and column or an error, if the number of rows or columns is
-    /// greater than the matrix or greater than the matrix.
-    ///
-    /// # Examples
-    /// ```rust
-    /// use qfall_math::integer_mod_q::MatZq;
-    /// use qfall_math::traits::GetEntry;
-    /// use qfall_math::integer::Z;
-    /// use std::str::FromStr;
-    ///
-    /// let matrix = MatZq::from_str("[[1, 2, 3],[4, 5, 6],[7, 8, 9]] mod 10").unwrap();
-    ///
-    /// assert_eq!(Z::from(3), matrix.get_entry(0, 2).unwrap());
-    /// assert_eq!(Z::from(8), matrix.get_entry(2, 1).unwrap());
-    /// assert_eq!(Z::from(8), matrix.get_entry(-1, -2).unwrap());
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type [`OutOfBounds`](MathError::OutOfBounds)
-    ///     if `row` or `column` are greater than the matrix size.
-    fn get_entry(
-        &self,
-        row: impl TryInto<i64> + Display,
-        column: impl TryInto<i64> + Display,
-    ) -> Result<Z, MathError> {
-        let (row_i64, column_i64) = evaluate_indices_for_matrix(self, row, column)?;
-
-        let mut out = Z::default();
-        let entry = unsafe { fmpz_mod_mat_entry(&self.matrix, row_i64, column_i64) };
-        unsafe { fmpz_set(&mut out.value, entry) };
-        Ok(out)
-    }
-}
-
-impl GetEntry<Zq> for MatZq {
-    /// Outputs the [`Zq`] value of a specific matrix entry.
-    ///
-    /// Parameters:
-    /// - `row`: specifies the row in which the entry is located
-    /// - `column`: specifies the column in which the entry is located
-    ///
-    /// Negative indices can be used to index from the back, e.g., `-1` for
-    /// the last element.
-    ///
-    /// Returns the [`Zq`] value of the matrix at the position of the given
-    /// row and column or an error, if the number of rows or columns is
-    /// greater than the matrix.
-    ///
-    /// # Examples
-    /// ```rust
-    /// use qfall_math::integer_mod_q::{MatZq, Zq};
-    /// use qfall_math::traits::GetEntry;
-    /// use std::str::FromStr;
-    ///
-    /// let matrix = MatZq::from_str("[[1, 2, 3],[4, 5, 6],[7, 8, 9]] mod 10").unwrap();
-    ///
-    /// assert_eq!(Zq::from((3, 10)), matrix.get_entry(0, 2).unwrap());
-    /// assert_eq!(Zq::from((8, 10)), matrix.get_entry(2, 1).unwrap());
-    /// assert_eq!(Zq::from((8, 10)), matrix.get_entry(-1, -2).unwrap());
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type [`OutOfBounds`](MathError::OutOfBounds)
-    ///     if `row` or `column` are greater than the matrix size.
-    fn get_entry(
-        &self,
-        row: impl TryInto<i64> + Display,
-        column: impl TryInto<i64> + Display,
-    ) -> Result<Zq, MathError> {
-        let value: Z = self.get_entry(row, column)?;
-
-        Ok(Zq::from((value, &self.modulus)))
-    }
-}
-
-impl MatZq {
     /// Outputs the row vector of the specified row.
     ///
     /// Parameters:
@@ -280,14 +185,14 @@ impl MatZq {
         let (row_1, col_1) = evaluate_indices_for_matrix(self, row_1, col_1)?;
         let (row_2, col_2) = evaluate_indices_for_matrix(self, row_2, col_2)?;
         assert!(
-            row_2 >= row_1,
-            "The number of rows must be positive, i.e. row_2 ({row_2}) must be greater or equal row_1 ({row_1})"
-        );
+                row_2 >= row_1,
+                "The number of rows must be positive, i.e. row_2 ({row_2}) must be greater or equal row_1 ({row_1})"
+            );
 
         assert!(
-            col_2 >= col_1,
-            "The number of columns must be positive, i.e. col_2 ({col_2}) must be greater or equal col_1 ({col_1})"
-        );
+                col_2 >= col_1,
+                "The number of columns must be positive, i.e. col_2 ({col_2}) must be greater or equal col_1 ({col_1})"
+            );
 
         // increase both values to have an inclusive capturing of the matrix entries
         let (row_2, col_2) = (row_2 + 1, col_2 + 1);
@@ -413,6 +318,124 @@ impl MatZq {
         }
 
         out
+    }
+}
+
+impl GetNumRows for MatZq {
+    /// Returns the number of rows of the matrix as a [`i64`].
+    ///
+    /// # Examples
+    /// ```
+    /// use qfall_math::integer_mod_q::MatZq;
+    /// use qfall_math::traits::*;
+    ///
+    /// let matrix = MatZq::new(5, 6, 7);
+    /// let rows = matrix.get_num_rows();
+    /// ```
+    fn get_num_rows(&self) -> i64 {
+        self.matrix.mat[0].r
+    }
+}
+
+impl GetNumColumns for MatZq {
+    /// Returns the number of columns of the matrix as a [`i64`].
+    ///
+    /// # Examples
+    /// ```
+    /// use qfall_math::integer_mod_q::MatZq;
+    /// use qfall_math::traits::*;
+    ///
+    /// let matrix = MatZq::new(5, 6, 7);
+    /// let rows = matrix.get_num_columns();
+    /// ```
+    fn get_num_columns(&self) -> i64 {
+        self.matrix.mat[0].c
+    }
+}
+
+impl GetEntry<Z> for MatZq {
+    /// Outputs the [`Z`] value of a specific matrix entry.
+    ///
+    /// Parameters:
+    /// - `row`: specifies the row in which the entry is located
+    /// - `column`: specifies the column in which the entry is located
+    ///
+    /// Negative indices can be used to index from the back, e.g., `-1` for
+    /// the last element.
+    ///
+    /// Returns the [`Z`] value of the matrix at the position of the given
+    /// row and column or an error, if the number of rows or columns is
+    /// greater than the matrix or greater than the matrix.
+    ///
+    /// # Examples
+    /// ```rust
+    /// use qfall_math::integer_mod_q::MatZq;
+    /// use qfall_math::traits::GetEntry;
+    /// use qfall_math::integer::Z;
+    /// use std::str::FromStr;
+    ///
+    /// let matrix = MatZq::from_str("[[1, 2, 3],[4, 5, 6],[7, 8, 9]] mod 10").unwrap();
+    ///
+    /// assert_eq!(Z::from(3), matrix.get_entry(0, 2).unwrap());
+    /// assert_eq!(Z::from(8), matrix.get_entry(2, 1).unwrap());
+    /// assert_eq!(Z::from(8), matrix.get_entry(-1, -2).unwrap());
+    /// ```
+    ///
+    /// # Errors and Failures
+    /// - Returns a [`MathError`] of type [`OutOfBounds`](MathError::OutOfBounds)
+    ///     if `row` or `column` are greater than the matrix size.
+    fn get_entry(
+        &self,
+        row: impl TryInto<i64> + Display,
+        column: impl TryInto<i64> + Display,
+    ) -> Result<Z, MathError> {
+        let (row_i64, column_i64) = evaluate_indices_for_matrix(self, row, column)?;
+
+        let mut out = Z::default();
+        let entry = unsafe { fmpz_mod_mat_entry(&self.matrix, row_i64, column_i64) };
+        unsafe { fmpz_set(&mut out.value, entry) };
+        Ok(out)
+    }
+}
+
+impl GetEntry<Zq> for MatZq {
+    /// Outputs the [`Zq`] value of a specific matrix entry.
+    ///
+    /// Parameters:
+    /// - `row`: specifies the row in which the entry is located
+    /// - `column`: specifies the column in which the entry is located
+    ///
+    /// Negative indices can be used to index from the back, e.g., `-1` for
+    /// the last element.
+    ///
+    /// Returns the [`Zq`] value of the matrix at the position of the given
+    /// row and column or an error, if the number of rows or columns is
+    /// greater than the matrix.
+    ///
+    /// # Examples
+    /// ```rust
+    /// use qfall_math::integer_mod_q::{MatZq, Zq};
+    /// use qfall_math::traits::GetEntry;
+    /// use std::str::FromStr;
+    ///
+    /// let matrix = MatZq::from_str("[[1, 2, 3],[4, 5, 6],[7, 8, 9]] mod 10").unwrap();
+    ///
+    /// assert_eq!(Zq::from((3, 10)), matrix.get_entry(0, 2).unwrap());
+    /// assert_eq!(Zq::from((8, 10)), matrix.get_entry(2, 1).unwrap());
+    /// assert_eq!(Zq::from((8, 10)), matrix.get_entry(-1, -2).unwrap());
+    /// ```
+    ///
+    /// # Errors and Failures
+    /// - Returns a [`MathError`] of type [`OutOfBounds`](MathError::OutOfBounds)
+    ///     if `row` or `column` are greater than the matrix size.
+    fn get_entry(
+        &self,
+        row: impl TryInto<i64> + Display,
+        column: impl TryInto<i64> + Display,
+    ) -> Result<Zq, MathError> {
+        let value: Z = self.get_entry(row, column)?;
+
+        Ok(Zq::from((value, &self.modulus)))
     }
 }
 
@@ -613,6 +636,39 @@ mod test_get_num {
         let matrix = MatZq::new(5, 10, 7);
 
         assert_eq!(matrix.get_num_columns(), 10);
+    }
+}
+
+#[cfg(test)]
+mod test_get_mat {
+    use crate::{
+        integer::Z,
+        integer_mod_q::MatZq,
+        traits::{GetEntry, GetNumColumns, GetNumRows, SetEntry},
+    };
+
+    /// Test if the dimensions are taken over correctly
+    #[test]
+    fn dimensions() {
+        let matzq = MatZq::new(15, 17, 13);
+
+        let matz_1 = matzq.get_mat();
+
+        assert_eq!(15, matz_1.get_num_rows());
+        assert_eq!(17, matz_1.get_num_columns());
+    }
+
+    /// Test if entries are taken over correctly
+    #[test]
+    fn entries_taken_over_correctly() {
+        let mut matzq = MatZq::new(2, 2, u64::MAX);
+        matzq.set_entry(0, 0, u64::MAX - 58).unwrap();
+        matzq.set_entry(0, 1, -1).unwrap();
+
+        let matz_1 = matzq.get_mat();
+
+        assert_eq!(Z::from(u64::MAX - 1), matz_1.get_entry(0, 1).unwrap());
+        assert_eq!(Z::from(u64::MAX - 58), matz_1.get_entry(0, 0).unwrap());
     }
 }
